data-prepper-expression/src/test/java/org/opensearch/dataprepper/expression/NotInOperatorTest.java
/*
 * Copyright OpenSearch Contributors
 * SPDX-License-Identifier: Apache-2.0
 */

package org.opensearch.dataprepper.expression;

import org.junit.jupiter.api.Test;

import java.util.Collections;
import java.util.Set;

import static org.hamcrest.CoreMatchers.is;
import static org.hamcrest.MatcherAssert.assertThat;
import static org.junit.jupiter.api.Assertions.assertThrows;

class NotInOperatorTest {
    final NotInOperator objectUnderTest = new NotInOperator();

    @Test
    void testGetSymbol() {
        assertThat(objectUnderTest.getSymbol(), is("not in"));
    }

    @Test
    void testEvalValidArgs() {
        assertThat(objectUnderTest.eval(1, Set.of(1)), is(false));
        assertThat(objectUnderTest.eval(1, Collections.emptySet()), is(true));
    }

    @Test
    void testEvalInValidArgLength() {
        assertThrows(IllegalArgumentException.class, () -> objectUnderTest.eval(1, 2, Set.of(1, 2)));
    }

    @Test
    void testEvalInValidArgType() {
        assertThrows(IllegalArgumentException.class, () -> objectUnderTest.eval(1, 1));
    }
}